add extra firebase helper functions
add functions to better test the registration process
